Added shellcheck
I have installed shellcheck which checks for possible issues in bash
script. In this commit I have changed all the warnings brought up by it,
and added to the script, an execution of shellcheck so that everytime
this script is executed it checks itself and quits if it fails in some
way.