Added function scripts
I added files that will contain all of the functions that the script
will contain. First I need to separate the script into functions, but
that is where they will go.

Shellcheck now checks all files with a .sh extension, so I added the .sh
extension to the end of all bash scripts, and I gave them shorter names.